Update entties and project properties.

+++ src/main/java/com/facsciences_planning_management/facsciences_planning_management/planning_service/managers/dtos/UeUpdateRequest.java
@@ -1,11 +1,11 @@
 package com.facsciences_planning_management.facsciences_planning_management.planning_service.managers.dtos;
 
 public record UeUpdateRequest(
-                String name,
-                String code,
-                Long credits,
-                String duration,
-                String category,
-                Integer hourlyCharge,
-                String levelId) {
+        String name,
+        String code,
+        Integer credits,
+        String duration,
+        String category,
+        Integer hourlyCharge,
+        String levelId) {
 }
